Use attachments rather than blocks
While these are described as legacy, I've opted to use attachments[1].
They are useful because they provide some colour output, which makes it
easy to spot failed builds from good builds.

I've chosen to roughly copy the output that we used for Jenkins[2],
although it's missing the time the build took. Not quite sure how to get
that yet, probably not too important at the moment.

[1] https://api.slack.com/reference/messaging/attachments#example
[2] https://github.com/futurelearn/jenkins-pipeline/blob/b8b5069eaf21968d69ce54e2724a8e02c8343249/vars/notify.groovy#L76-L100