Use git bundle rather than archive
When testing these backups I realised that archive just keeps the tree
of the latest commit, rather than keeping the commit history. The commit
history is as important as the files themselves, so instead let's use
`git bundle`[1].

To restore a bundle you can clone the repository from the bundle:

```
git clone bundle_file /destination/directory
```

I've opted to put `git-bundle` in the name of the backup to make it
clear exactly what format it is, but will of course also write
documentation and a process for recovery.

[1] https://www.git-scm.com/docs/git-bundle