Fix issue with notifications
With a more complicated pipeline, you cannot use `job.status` to get the
status of the full workflow, because it only reports the single job
you're running it on. If you use the `needs` conditional, then it's
going to give you inaccurate results.

Instead, what we can do is explicitly declare a success on success, and
a failure on failure.