Escape double quotes
This caused an issue where double quotes are used in commit messages,
they were not escaped and caused the payload sent to the API to be
incorrect. This escapes the double quotes using `sed`.

I guess there might be other places we need to escape strings... but now
we output the payload, it's easy to spot!